Border
The border is drawn between the margin and the padding of a box. If you are using the standard box model, the size of the border is added to the width and height of the content box. If you are using the alternative box model, then the bigger the border is, the smaller the content box is, as the border takes up some of that available width and height of the element box

    For styling borders, there are a large number of properties — there are four borders, and each border has a style, width, and color that we might want to manipulate